requirements.txt
allure-python-commons~=2.15.0
pytest~=8.4.1
requests~=2.32.4
jsonpath-ng~=1.7.0
PyYAML~=6.0.2
pandas~=2.3.1
PyMySQL~=1.1.1